requirements.txt
ollama>=0.1.7
pyttsx3>=2.90
SpeechRecognition>=3.10.0
PyAudio>=0.2.13
pyperclip>=1.8.2
psutil>=5.9.0
requests>=2.31.0
beautifulsoup4>=4.12.0
rich>=13.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
keyboard>=0.13.5
pyautogui>=0.9.54
plyer>=2.1.0
chromadb>=0.4.0
numpy>=1.24.0
